refactor: 용도에 맞게 패키지 분리
