+++ common/src/main/java/store/_0982/common/domain/sellerpayout/SellerPayoutFailure.java
@@ -1,4 +1,4 @@
-package store._0982.common.domain.settlement;
+package store._0982.common.domain.sellerpayout;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
