+++ common/src/main/java/store/_0982/common/domain/sellerpayout/SellerPayoutStatus.java
@@ -1,4 +1,4 @@
-package store._0982.common.domain.settlement;
+package store._0982.common.domain.sellerpayout;
 
 public enum SellerPayoutStatus {
     PENDING,
